Move AoA->_rasterize to Vector->rasterize, with the following API changes:
* add mandatory min/max args.
* return undef to indicate loop end.

Fix up the callers to support the new API.

This function allows the caller to enumerate over every point in an arbitrary
N-dimensional space without recursion.  The space is specified by the min/max
arguments.  Rasterize is very useful for the "generic" storage classes.

It is also a good candidate for XS acceleration.  More on that later.