Create a _copy method for LBSGV.
Copying the Vec storage object is easy, because a deep copy is not
necessary.  I've made this a private method, because copying other
types of Storage objects will not be nearly as trivial, and there
may not be any benefit to doing so.

Creating a 4-dimensional torus is very expensive - being able to
cache and reuse them gives me huge speedups (600% for CPU-intensive
files, 400% for my test suite as a whole).