Fix stdio_in_num() and stdio_in_ascii()
* they need to reflect when an EOF is encountered.
* they should not store input overflow in the IP, because that
  breaks concurrent access.

I am moving the stdin handling into the Interpreter module, so the
overflow is accessible to all IPs.  I am also using sysread to only
read one byte at a time, to minimize issues with multiple
interpreters running in the same process.

(I don't think it is possible to eliminate such issues, because
stdio_in_num() needs to read ahead to find the first non-digit,
and that extra character still needs to be stored somewhere.)