Test lib/TEST.pm properly.  This brings total coverage to 100%.
I could not test it all with a single test script, I had to use 2.
Turns out the perl Test.pm and the Test::Builder instance created
by befunge's TEST.pm share the same plan, and a redundant plan
results in an error message.  Thus, I couldn't test both sides of
the no_plan condition in the plan function, and had to use two
test scripts to cover that.