fix repeat shenanigan (yet again)
see mycology readme:
==== README ====
The specification does not say that the operand should be skipped over
after execution. The only special case is when the amount of times to
execute is zero.

This means that 2k6 should indeed first push 2 sixes at the k, and then
a third when encountering the 6 itself.

This also means that there is no way to execute an instruction only
once: 1k6 results in two sixes. (Another IP may certainly modify the 6
immediately after the k is executed, but that's a somewhat unlikely case
and not exactly a good way to handle this limitation.)

The spec is somewhat unclear on the entirety of k, but both of the above
issues have been confirmed with Chris Pressey.
==== /README ====